+++ backend/src/daoBean/DrzavaDaoLocal.java
@@ -3,6 +3,6 @@ package daoBean;
 import model.Drzava;
 import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDao;
 
-public interface DrzavaDaoLocal extends GenericDao<Drzava, Long> {
+public interface DrzavaDaoLocal extends GenericDao<Drzava, Integer> {
 
 }
