+++ backend/src/daoBean/MestoDao.java
@@ -8,6 +8,6 @@ import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDaoBean;
 
 @Stateless
 @Local(MestoDaoLocal.class)
-public class MestoDao extends GenericDaoBean<Mesto, Long> implements MestoDaoLocal{
+public class MestoDao extends GenericDaoBean<Mesto, Integer> implements MestoDaoLocal{
 
 }
