+++ backend/src/daoBean/MestoDaoLocal.java
@@ -3,6 +3,6 @@ package daoBean;
 import model.Mesto;
 import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDao;
 
-public interface MestoDaoLocal extends GenericDao<Mesto, Long>{
+public interface MestoDaoLocal extends GenericDao<Mesto, Integer>{
 
 }
