+++ backend/src/daoBean/PoslovnaGodinaDaoLocal.java
@@ -3,6 +3,6 @@ package daoBean;
 import model.PoslovnaGodina;
 import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDao;
 
-public interface PoslovnaGodinaDaoLocal extends GenericDao<PoslovnaGodina, Long> {
+public interface PoslovnaGodinaDaoLocal extends GenericDao<PoslovnaGodina, Integer> {
 
 }
