+++ backend/src/daoBean/PreduzeceDaoLocal.java
@@ -3,6 +3,6 @@ package daoBean;
 import model.Preduzece;
 import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDao;
 
-public interface PreduzeceDaoLocal extends GenericDao<Preduzece, Long> {
+public interface PreduzeceDaoLocal extends GenericDao<Preduzece, Integer> {
 
 }
